### 2018-03-14 #619 @zedtaha
@leeper as reviewer
