### 2018-03-14 #619 @leeper
@zedtaha Would you be interested in reviewing this package for JOSS?
