### 2018-03-14 #619 @zedtaha
Hi @leeper do you mean nardl r package ? if yes I'm the developer!
