### 2018-03-14 #623 @whedon
[REVIEW]: nardl:An R package to estimate the nonlinear cointegrating autoregressive distributed lag model
**Submitting author:** @zedtaha (<a href="http://orcid.org/0000-0002-4488-5774">zaghdoudi Taha</a>)
**Repository:** <a href="https://github.com/zedtaha/nardl" target ="_blank">https://github.com/zedtaha/nardl</a>
**Version:** 0.1.3
