### 2018-03-27 5db07af
Added test structure to package
