### 2018-03-27 9a47b74
Added test structure to package
