### 2018-03-27 f394964
Added testthat test suite
