+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(nardl)
+
+test_check("nardl")
