+++ DESCRIPTION
@@ -10,3 +10,4 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
 Imports: stats, strucchange, tseries, matlab, Formula, gtools
+Suggests: testthat
