### 2018-03-27 f3bb668
Separated fod doc and corrected itemize
