+++ R/fod.R
@@ -1,0 +1,14 @@
+#' Indian yearly data of inflation rate and percentage food import to total import
+#'
+#' The data frame \code{fod} contains the following variables:
+#' \itemize{
+#' \item{food: }{percentage food import to total import}
+#' \item{inf: }{inflation rate}
+#' \item{year: }{the year}
+#' }
+#'
+#' @docType data
+#' @name  fod
+#' @usage data(fod)
+#' @format A data frame with 54 rows and 2 variables
+"fod"
