### 2018-03-27 5b5752f
Restructured code files: move helper functions nardl to own file.
