### 2018-03-27 67ba5ce
Slight improvement trimr function: use matrix() instead of cbind()
