### 2018-03-27 e8e58bd
Optimezed lagm: get rid of double loop and growing object.
