### 2018-03-27 8d21aa2
Adapted nardl to work correctly with lagm and removed trimr
