### 2018-03-27 ef7f526
Removed ginv
