+++ R/helperfunctions_nardl.R
@@ -84,21 +84,6 @@ ArchTest <- function (x, lags=12, demean = FALSE)
   return(result)
 }
 
-#-------------------------------------------------------------------------------
-# Function ginv
-
-# This function is not used anywhere
-ginv <- function(x, tol = sqrt(.Machine$double.eps))
-{
-  ## Generalized Inverse of a Matrix
-  dnx <- dimnames(x)
-  if(is.null(dnx)) dnx <- vector("list", 2)
-  s <- svd(x)
-  nz <- s$d > tol * s$d[1]
-  structure(
-    if(any(nz)) s$v[, nz] %*% (t(s$u[, nz])/s$d[nz]) else x,
-    dimnames = dnx[2:1])
-}
 
 #-------------------------------------------------------------------------------
 # Function bp2
