### 2018-03-27 ca62f65
Added test for helper functions
