### 2018-03-27 35a025e
Made 3 the default value for case in nardl() and removed redundant code
