### 2018-03-27 18a16cb
Removed logical == TRUE construct (that's redundant)
