### 2018-03-27 e170a35
Moved cumsq, cusum and seqa to helper file
