### 2018-03-27 zedtaha/nardl#3 @JoFAM
Dataset fod documentation issues
- [ ] The dataset fod needs either a reference to an article where it comes from or another mention of its origin (own research, simulated, grabbed from database X at date Y, ...)
- [x] The dataset documentation best goes in its own file for easy maintenance. This is changed in commit https://github.com/JoFAM/nardl/commit/f3bb6681ff7976894ce682268d01bbcac327aecd
- [x] The layout of the variable explanation is messed up (used `\itemize` with two inputs for `\item`. If you want to use that construct, you need `\describe` instead of `\itemize`). This is changed in commit https://github.com/JoFAM/nardl/commit/f3bb6681ff7976894ce682268d01bbcac327aecd
