### 2018-03-27 zedtaha/nardl#5 @JoFAM
Create unit testing for checks and maintenance
It's good practice to add at least some unit testing so you can check whether all (user) functions give a correct result. It also helps when refactoring code or extending the package to make sure you don't accidently break anything.

I've set this up with `testthat` in commit : https://github.com/JoFAM/nardl/commit/f394964b93a20a2130e2541fa684c0379e6eb47e

Tests for lagm are added in commit : https://github.com/JoFAM/nardl/commit/ca62f65ab8f7a06071860fca4a87fa7357342329

If you're not familiar with the `testthat` package and how to use it, this is a good start:

http://r-pkgs.had.co.nz/tests.html
