### 2018-03-28 zedtaha/nardl#8 @JoFAM
Help files nardl and summary.nardl incomplete
Help file nardl needs a basic description and a return value at least.

Help file summary.nardl doesn't exist, but it is an exported function.
