+++ man/summary.nardl.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/nardl.R
+% Please edit documentation in R/summary.nardl.R
 \name{summary.nardl}
 \alias{summary.nardl}
-\title{summary}
+\title{Summary of a nardl model}
 \usage{
 \method{summary}{nardl}(object, ...)
 }
@@ -11,6 +11,10 @@
 
 \item{...}{not used}
 }
+\value{
+an object of the S3 class \code{summary.nardl} with the
+following components:
+}
 \description{
-summary
+\code{summary} method for a \code{\link{nardl}} model.
 }
