### 2018-03-28 zedtaha/nardl#12 @JoFAM
Guidelines: at least link for bug reports and support needed
This should be at least included in DESCRIPTION, and also updated in the readme.md of the github repo so it is clear how to file bug reports or get support.
