### 2018-03-28 zedtaha/nardl#12 @JoFAM
Solved in commit https://github.com/JoFAM/nardl/commit/b974c61ce9c84a5a83d3cd6661ac529ed21d0cd4
