### 2018-03-28 #623 @JoFAM
The package provides new functionality for R: nonlinear ARDL models currently have to be built manually or calculated using other software.

## Current issues that warrant "major revision":

- There is no guidelines for bug reports or contributions : https://github.com/zedtaha/nardl/issues/12
- There is no testing done, results are impossible to check: https://github.com/zedtaha/nardl/issues/5
- help files for the few exported functions don't contain return values etc. : https://github.com/zedtaha/nardl/issues/8

## General remarks

In general the code is very ad-hoc and doesn't allow to easily extract test results for the different tests provided. None of the test functions are exported, so they can't be called directly. There's no real API to work with the S3 model object returned by nardl(). There's also no documentation as to what can be found in that object.

In order to be useful in a real world analysis, the API needs to be extended so one can at least work with the returned objects. In order to be maintainable, the code needs serious refactoring so it is more modularized.

@leeper : this is my assessment of the package nardl at its current state.
