### 2018-03-28 #623 @leeper
@JoFAM Thank you very much for your review! This is very useful feedback. Much appreciated.

@zedtaha I think the review does a very good job of identifying weaknesses in the structure of the code that you could address, particularly with regarding to making a fully featured API. Please take a look at the open issues and, especially, address https://github.com/zedtaha/nardl/issues/5. Thanks!
