### 2018-03-28 #623 @JoFAM
@zedtaha If you want, I can create a pull request from my fork, so you have already some of the issues (partly) taken care of. Also if you want further assistance with these and other issues, feel free to ping me. I'll be glad to help you clean up your already very valuable work.
