### 2018-03-28 #623 @JoFAM
@zedtaha Super. Do I create a pull request directly on your master branch, or do you want to create a devel branch first?
