+++ R/summary.nardl.R
@@ -25,13 +25,8 @@ summary.nardl<-function(object,...)
   cat("---------------------------------\n")
   cat("\n model diagnostic tests:\n----------\n")
   cat(" JB test:\n","JB:",object$jbtest$statistic[[1]],"Pvalue",object$jbtest$p.value[[1]],"\n----------\n")
-  #cat(" LM test for serial correlation:\n","LM(",object$np,"):",object$lmtest$statistic[[1]],"Pvalue",object$lmtest$p.value[[1]],"\n----------\n")
   cat(" LM test for serial correlation:\n","LM(",object$np,"):",object$lm2[1],"Pvalue",object$lm2[2],"\n----------\n")
   cat(" ARCH test:\n","ARCH(",object$np,"):",object$arch$statistic[[1]],"Pvalue",object$arch$p.value[[1]],"\n----------\n")
-  cat("==============================================================\n")
-  cat("Asymmetric Cointegration test\n")
-  # print( bounds.test(3,object$k,object$fstat))
-  pssbounds(case=object$case,obs=object$obs,fstat=object$fstat,k = object$k)
   cat("==============================================================\n")
   cat("\nLong-run coefficients\n")
   printCoefmat(object$lres,has.Pvalue = TRUE,signif.stars = TRUE)
