+++ NAMESPACE
@@ -1,8 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(summary,nardl)
+export(ArchTest)
+export(bp2)
 export(nardl)
 export(plotmplier)
+export(pssbounds)
 import(Formula)
 import(matlab)
 importFrom(graphics,abline)

[…]
