### 2018-04-19 4daead7
Update paper.md
