### 2018-04-19 #623 @zedtaha
Hi @JoFAM i made some changes in the nardl package  and i invite you to take a look
