### 2018-04-19 #623 @zedtaha
@JoFAM hope that the changes meet up all the issues ;)
