+++ _config.yml
@@ -1,1 +1,3 @@
 theme: jekyll-theme-cayman
+title: NARDL R package
+description: Nonlinear autoregrissive distributes lagged model
