+++ index.md
@@ -5,7 +5,10 @@ output: github_document
 [![](http://www.r-pkg.org/badges/version/nardl)]()
 [![](http://cranlogs.r-pkg.org/badges/nardl)](http://cran.rstudio.com/web/packages/nardl/index.html)
 # nardl
-nardl:An R package to estimate the nonlinear cointegrating autoregressive distributed lag model
+
+**nardl: An R package to estimate the nonlinear cointegrating autoregressive distributed lag model**
+
+**Author : Dr. Taha Zaghdoudi**
 
 In this package, we apply the ordinary least squares method to estimate the cointegrating nonlinear ARDL (NARDL) model in which short and long-run nonlinearities are introduced via positive and negative partial sum decompositions of the explanatory variables.Besides, we provide the CUSUM, CUSUMSQ model stability tests, model selection via aic, bic and rsqaured criteria and the dynamic multipliers plot.
 
