### 2018-04-25 bd5e6db
Update dynmplier.R
