+++ Paper/paper.bib
@@ -10,6 +10,7 @@
 @Manual{ztahanardl,
     title = {nardl: Nonlinear Cointegrating  Autoregressive Distributed Lag Model},
     author = {Taha, Zaghdoudi},
-    note = {R package version 0.1.2},
-	year={2018}
+    note = {R package version 0.1.4},
+	year={2018},
+	doi = {10.5281/zenodo.1221245}
   }
