+++ man/cumsq.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helperfunctions_nardl.R
+\name{cumsq}
+\alias{cumsq}
+\title{Function cumsq}
+\usage{
+cumsq(e, k, n)
+}
+\arguments{
+\item{e}{is the recursive errors}
+
+\item{k}{is the estimated coefficients length}
+
+\item{n}{is the recursive errors length}
+}
+\description{
+Function cumsq
+}
+\examples{
+
+reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
+e<-reg$rece
+k<-reg$k
+n<-reg$n
+cumsq(e=e,k=k,n=n)
+
+}
