+++ man/cusum.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helperfunctions_nardl.R
+\name{cusum}
+\alias{cusum}
+\title{Function cusum}
+\usage{
+cusum(e, k, n)
+}
+\arguments{
+\item{e}{is the recursive errors}
+
+\item{k}{is the estimated coefficients length}
+
+\item{n}{is the recursive errors length}
+}
+\description{
+Function cusum
+}
+\examples{
+
+reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
+e<-reg$rece
+k<-reg$k
+n<-reg$n
+cusum(e=e,k=k,n=n)
+
+}
