+++ NAMESPACE
@@ -3,11 +3,12 @@
 S3method(summary,nardl)
 export(ArchTest)
 export(bp2)
+export(cumsq)
+export(cusum)
 export(nardl)
 export(plotmplier)
 export(pssbounds)
 import(Formula)
-import(matlab)
 importFrom(graphics,abline)
 importFrom(graphics,legend)
 importFrom(graphics,lines)

[…]
