### 2018-05-22 zedtaha/nardl#15 @JoFAM
Failing unittest Case III
Unit test gives following error when running R CMD check

Case III more than 45 obs gives correct results
pssbounds(...) not equal to `pssoutput`.
target is NULL, current is list
