### 2018-05-22 3aa6265
Made cusum a proper S3 method.
