+++ man/cusum.Rd
@@ -1,27 +1,39 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helperfunctions_nardl.R
+% Please edit documentation in R/cusum.R
 \name{cusum}
 \alias{cusum}
+\alias{cusum.nardl}
+\alias{cusum.default}
 \title{Function cusum}
 \usage{
-cusum(e, k, n)
+cusum(x, ...)
+
+\method{cusum}{nardl}(x, ...)
+
+\method{cusum}{default}(x, k, n, ...)
 }
 \arguments{
-\item{e}{is the recursive errors}
+\item{x}{is a object of the class \code{\link{nardl}} or a vector
+with the recursive errors}
+
+\item{...}{arguments passed to other methods. Currently ignored.}
 
 \item{k}{is the estimated coefficients length}
 
 \item{n}{is the recursive errors length}
 }
 \description{
-Function cusum
+This function creates two plots for the cusum test.
 }
 \examples{
 
 reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
+cusum(reg)
+
+# Or by hand
 e<-reg$rece
 k<-reg$k
 n<-reg$n
-cusum(e=e,k=k,n=n)
+cusum(x=e,k=k,n=n)
 
 }
