### 2018-05-22 bb4977b
Created proper S3 method for cumsq
