+++ man/cumsq.Rd
@@ -1,13 +1,22 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helperfunctions_nardl.R
+% Please edit documentation in R/cumsq.R
 \name{cumsq}
 \alias{cumsq}
+\alias{cumsq.nardl}
+\alias{cumsq.default}
 \title{Function cumsq}
 \usage{
-cumsq(e, k, n)
+cumsq(x, ...)
+
+\method{cumsq}{nardl}(x, ...)
+
+\method{cumsq}{default}(x, k, n, ...)
 }
 \arguments{
-\item{e}{is the recursive errors}
+\item{x}{is a object of the class \code{\link{nardl}} or a vector
+with the recursive errors}
+
+\item{...}{arguments passed to other methods. Currently ignored.}
 
 \item{k}{is the estimated coefficients length}
 
@@ -19,9 +28,12 @@ Function cumsq
 \examples{
 
 reg<-nardl(food~inf,fod,ic="aic",maxlags = TRUE,graph = TRUE,case=3)
+cumsq(reg)
+
+# Or manually
 e<-reg$rece
 k<-reg$k
 n<-reg$n
-cumsq(e=e,k=k,n=n)
+cumsq(x=e,k=k,n=n)
 
 }
