+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(cumsq,default)
+S3method(cumsq,nardl)
 S3method(cusum,default)
 S3method(cusum,nardl)
 S3method(summary,nardl)
