### 2018-05-22 zedtaha/nardl#5 @JoFAM
@zedtaha There's an error in the unit tests (see also https://github.com/zedtaha/nardl/issues/15)
