### 2018-05-22 b80d17d
make summary.nardl create object and add a print method for it
