+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(cumsq,default)
 S3method(cumsq,nardl)
 S3method(cusum,default)
 S3method(cusum,nardl)
+S3method(print,summary.nardl)
 S3method(summary,nardl)
 export(ArchTest)
 export(bp2)
