### 2018-05-22 zedtaha/nardl#8 @JoFAM
summary.nardl now has a components list in commit  b80d17d

@zedtaha you can use that one as an example of what the help page of the nardl package needs. As nardl() is your main function, that's one thing you really should document extensively. Otherwise nobody knows where to get what information to eg use your pkg in simulations etc.
