### 2018-05-22 e88e5b7
updated help file summary.nardl : np turns out to be the number of lags
