+++ man/summary.nardl.Rd
@@ -18,7 +18,7 @@ following components:
  \itemize{
    \item icres: a matrix with the AIC, BIC, LL and R2 values
    \item sel: a \code{\link[stats]{summary.lm}} object with the actual fit
-   \item np: the number of p
+   \item np: the number of lags
    \item jbtest: a list with the results of the JB test
    \item lmtest: a list with the statistic nd p value related to
    the LM test.
