### 2018-05-22 fecd7dd
Added ArchTest methods so they work on nardl objects
