+++ NAMESPACE
@@ -1,5 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(ArchTest,default)
+S3method(ArchTest,nardl)
+S3method(ArchTest,summary.nardl)
 S3method(cumsq,default)
 S3method(cumsq,nardl)
 S3method(cusum,default)
