### 2018-05-22 9db8dc5
Added lmtest (bp2) to the functions with appropriate methods for nardl and summary.nardl
