+++ NAMESPACE
@@ -3,6 +3,8 @@
 S3method(ArchTest,default)
 S3method(ArchTest,nardl)
 S3method(ArchTest,summary.nardl)
+S3method(LMTest,nardl)
+S3method(LMTest,summary.nardl)
 S3method(cumsq,default)
 S3method(cumsq,nardl)
 S3method(cusum,default)
@@ -10,6 +12,7 @@ S3method(cusum,nardl)
 S3method(print,summary.nardl)
 S3method(summary,nardl)
 export(ArchTest)
+export(LMTest)
 export(bp2)
 export(cumsq)
 export(cusum)
