### 2018-05-22 c549e95
Added ... argument to documentation LMTest
