### 2018-05-22 440bc3c
Upped version to 0.1.6
