+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: nardl
 Type: Package
 Title: Nonlinear Cointegrating  Autoregressive Distributed Lag Model
-Version: 0.1.5
+Version: 0.1.6
 Author: Taha Zaghdoudi
 Maintainer: Taha Zaghdoudi <zedtaha@gmail.com>
 Description: Computes the nonlinear cointegrating autoregressive distributed lag model with p lags of the dependent variables and q lags of independent variables proposed by (Shin, Yu & Greenwood-Nimmo, 2014 <doi:10.1007/978-1-4899-8008-3_9>).
