### 2018-05-22 #623 @JoFAM
@zedtaha I have another pull request that tackles a number of the issues from the previous review round:

https://github.com/zedtaha/nardl/pull/17

## General remarks:

The package improved in this round, but still doesn't provide basic functionality one would expect. At least one should be able to extract coefficients and other important model statistics from a model object. They might need these to plot, to construct their own tests, to do simulation  studies, ...

The documentation of the returned objects can be more helpful. Eg, it took me quite a while to realize 'np' is actually the number of lags, what 'sel' and 'fit' are and so forth.

There is much improvement possible in the unit testing. Currently the code coverage as reported by the `covr` package is about 14%. Even when the plot functions are ignored, there's not a single test for the pssbounds function or the summary.nardl function.

## Points on which improvement is necessary

For me, the minimal requirements to give this package a pass, are the following :

- Make sure pssbounds actually outputs something. Now the usefulness of this function is very limited: there's no easy way the results of that function can be captured and used in other code.
- Add a few more unit tests for the important functions of the package. A code coverage of 13% is very low. Especially the pssbounds function with its 1700 lines really needs some testing.
- Solve the failing unit test ( issue https://github.com/zedtaha/nardl/issues/15 ). That is actually a placeholder and example I added in the previous PR to get you started.
- the output of the function nardl needs describing ( issue https://github.com/zedtaha/nardl/issues/8 )
- You really need to give at least some indication of the source of the data ( issue https://github.com/zedtaha/nardl/issues/3 )

@leeper : This is my assesment of the package nardl at its current state.
