### 2018-05-23 #623 @leeper
@JoFAM Thanks! This looks great. Really, really helpful review.

@zedtaha Can you incorporate the PR and try to address the remaining issues? Given the scale of @JoFAM's contribution via the PR, you may also want to list him in DESCRIPTION as a contributor.
