### 2018-08-08 #623 @leeper
@zedtaha Just another nudge to complete the requested revisions.
