### 2018-10-14 #623 @zedtaha
@leeper I apologize for the delay I was on a working trip. However, there is so many issues to do so I think its correct to withdrew the paper. Best
