### 2018-10-14 #623 @zedtaha
@leeper I'm also open to more suggestion.
