### 2019-03-02 #623 @faheemja
i have a times series data matrix i want to fit functional autoregessive model of order one and two in R..how i can do do this in R???
