### 2019-03-02 #623 @danielskatz
@faheemja - this thread is was a review for a paper for the Journal of Open Source software, and is closed.

The repository for the software that was being reviewed, where you might want to ask your question, is https://github.com/zedtaha/nardl
